// types/species.ts
export interface Species {
  id: string;
  name: string;
  englishName: string;
  scientificName: string;
  region: string;
  family: string;
  order: string;
  quantity: string;
  commonWeight: string;
  locationDetails: string;
  fishingTechnique: string;
  photoSource: string;
  images: string[];
  hasImages: boolean;
}